+++ docs/changelog.rst
@@ -17,6 +17,7 @@ New
 
 Changed
 -------
+- ``open_ncs`` returns either ``dict[Path, xr.Dataset]`` or ``dict[str, xr.Dataset]``. When the uri contains no ``{name}`` placeholder, the full paths are used as keys. If there is a placehodler the key is the value of the ``{name}`` placeholder.
 -  Instead of skipping the ``global`` section by default, ``read_workflow_yaml`` now tries to convert all strings in all sections into absolute paths by default, by checking if the absolute path exists.
 
 Fixed